% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_how_you.R
\docType{data}
\name{hi_how_you}
\alias{hi_how_you}
\title{\code{hi_how_you} dataset}
\format{
A data frame with 3 variables:
Name- The student's name.
Math- The Math score.
Science- The Science score.
}
\usage{
hi_how_you
}
\description{
This data frame contains information about students' names, Math scores, and Science scores.
}
\examples{
data(hi_how_you)
}
\keyword{datasets}
